+++ src/database.ts
@@ -1,31 +1,22 @@
 import { AsyncLocalStorage } from 'async_hooks'
-import { getLogger } from 'log4js'
-import { ClientSession, Collection, Db, Logger, MongoClient, MongoClientCommonOption, SessionOptions } from 'mongodb'
+import { ClientSession, Collection, Db, MongoClient, MongoClientCommonOption, MongoClientOptions, SessionOptions } from 'mongodb'
 
 const sessionStorage = new AsyncLocalStorage<ClientSession>()
 
 export interface IMongo {
-  connect: () => void
+  connect: () => Promise<MongoClient>
   database: (dbname?: string, options?: MongoClientCommonOption) => Promise<Db>
   session: (options?: SessionOptions) => Promise<ClientSession>
   withTransaction: <T>(callback: () => Promise<T>, options?: SessionOptions) => Promise<T>
 }
 
-const logger = getLogger('db')
-
-export const Mongo = (url: string, db: string): IMongo => {
-  const client = new MongoClient(url, {
-    useUnifiedTopology: true
-  })
-  Logger.setLevel('debug')
-  Logger.filter('class', ['Cursor', 'Db'])
-  Logger.setCurrentLogger(function (msg, state) {
-    logger.debug(state?.className, state?.pid, state?.message)
-  })
+export const Mongo = (url: string, db: string, options: MongoClientOptions = { useUnifiedTopology: true }): IMongo => {
+  const client = new MongoClient(url, options)
   return {
     connect: async () => {
       await client.connect()
-      logger.info('🔗 Connected to Mongo')
+      console.debug('🔗 Connected to Mongo')
+      return client
     },
     database: async (dbname, options) => {
       return client.db(dbname ?? db, options)
@@ -37,18 +28,18 @@ export const Mongo = (url: string, db: string): IMongo => {
       const parentSession = sessionStorage.getStore()
       if (parentSession !== undefined) {
         const session = client.startSession(options)
-        logger.debug('starting session')
+        console.debug('starting session')
         return await sessionStorage.run<Promise<T>>(session, async () => {
           session.startTransaction()
-          logger.debug('starting transaction')
+          console.debug('starting transaction')
           try {
             const ret = await callback()
             await session.commitTransaction()
-            logger.debug('commit transaction')
+            console.debug('commit transaction')
             return ret
           } catch (e) {
             await session.abortTransaction()
-            logger.debug('rollback transaction')
+            console.debug('rollback transaction')
             throw e
           }
         })
